refactor(core/validation): Move ValidationMessage from core/validation to core/presentation/forms/validation
