+++ app/scripts/modules/core/src/index.ts
@@ -85,8 +85,6 @@ export * from './task';
 
 export * from './utils';
 
-export * from './validation';
-
 export * from './widgets';
 
 export * from './yamlEditor';
