+++ app/scripts/modules/core/src/presentation/forms/validation/index.ts
@@ -1,6 +1,7 @@
-export * from './categories';
-export * from './useValidationData';
-export * from './validators';
-export * from './validation';
 export * from './FormValidator';
 export * from './FormValidatorField';
+export * from './ValidationMessage';
+export * from './categories';
+export * from './useValidationData';
+export * from './validation';
+export * from './validators';
