--- app/scripts/modules/core/src/validation/index.ts
@@ -1,1 +1,0 @@
-export * from './ValidationMessage';
